+++ app/Http/Controllers/LiabilitySubSubCategoryController.php
@@ -14,6 +14,10 @@ class LiabilitySubSubCategoryController extends Controller
      */
     public function index()
     {
+        // Check if the user has permission to access this page
+        if (auth()->user()->access->liability == 3) {
+            return redirect()->route('admin.dashboard')->with('error', 'You do not have permission to access this page.');
+        }
         // Fetch all liability sub-subcategories from the database
         $liabilitySubSubCategories = LiabilitySubSubCategory::all();
 
@@ -37,14 +41,30 @@ class LiabilitySubSubCategoryController extends Controller
      */
     public function store(Request $request)
     {
+        // Check if the user has permission to create liability sub-subcategories
+        if (auth()->user()->access->liability != 2) {
+            return redirect()->route('admin.dashboard')->with('error', 'You do not have permission.');
+        }
         // Validate the request data
         $request->validate([
-            'name' => 'required|string|max:255|unique:liability_sub_sub_categories,name',
+            'name' => 'required|string|max:255',
             'description' => 'nullable|string|max:1000',
             'liability_sub_category_id' => 'required|exists:liability_sub_categories,id',
             'liability_category_id' => 'required|exists:liability_categories,id',
+            'slug' => 'required|string|max:255',
         ]);
 
+        $baseSlug = $request->slug;
+        $slug = $baseSlug;
+        $counter = 1;
+
+        // Check if slug exists in the contacts table
+        while (LiabilitySubSubCategory::where('slug', $slug)->exists()) {
+            $slug = $baseSlug . '-' . $counter++;
+        }
+
+        $data['slug'] = $slug;
+        $request->merge(['slug' => $data['slug']]);
         // Create a new liability sub-subcategory
         LiabilitySubSubCategory::create($request->all());
 
@@ -75,16 +95,30 @@ class LiabilitySubSubCategoryController extends Controller
      */
     public function update(Request $request, string $id)
     {
+        // Check if the user has permission to update liability sub-subcategories
+        if (auth()->user()->access->liability != 2) {
+            return redirect()->route('admin.dashboard')->with('error', 'You do not have permission.');
+        }
         // Validate the request data
         $request->validate([
-            'name' => 'required|string|max:255|unique:liability_sub_sub_categories,name,' . $id,
+            'name' => 'required|string|max:255',
             'description' => 'nullable|string|max:1000',
             'liability_sub_category_id' => 'required|exists:liability_sub_categories,id',
             'liability_category_id' => 'required|exists:liability_categories,id',
+            'slug' => 'required|string|max:255',
         ]);
 
         // Update the specified liability sub-subcategory
         $liabilitySubSubCategory = LiabilitySubSubCategory::findOrFail($id);
+        $baseSlug = $request->slug;
+        $slug = $baseSlug;
+        $counter = 1;
+        // Check if slug exists in the contacts table
+        while (LiabilitySubSubCategory::where('slug', $slug)->where('id', '!=', $id)->exists()) {
+            $slug = $baseSlug . '-' . $counter++;
+        }
+        $data['slug'] = $slug;
+        $request->merge(['slug' => $data['slug']]);
         $liabilitySubSubCategory->update($request->all());
 
         // Redirect back to the index with a success message
@@ -99,6 +133,10 @@ class LiabilitySubSubCategoryController extends Controller
      */
     public function destroy(string $id)
     {
+        // Check if the user has permission to delete liability sub-subcategories
+        if (auth()->user()->access->liability != 2) {
+            return redirect()->route('admin.dashboard')->with('error', 'You do not have permission.');
+        }
         // Find the liability sub-subcategory by ID and delete it
         $liabilitySubSubCategory = LiabilitySubSubCategory::findOrFail($id);
         $liabilitySubSubCategory->delete();
